UPSTREAM: mb/google/dedede: Add GPIO list
Leave all the GPIOs in not connected state so that they can be
configured depending on the use-case. This is done to park the GPIOs in
a known safe state. This will also help to ensure that the required
GPIOs are configured when the concerned use-cases are enabled.

Below GPIOs are configured in Native Function 1 and are required for
boot-up.
* VCCIN_AUX_VID0
* VCCIN_AUX_VID1
* AP_SLP_S0_L
* PLT_RST_L
* CPU_C10_GATE_L
* GPDs

BUG=None
TEST=Build and boot the mainboard.

Change-Id: I6a904ab1e06c4f9592040e33fcc4fe3b68242866
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8499f7fb1b961f2d3bf1329897d0c6ff97ab5038
Original-Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Change-Id: I5293536f66a6b08c9c2d2a6281684755a0c0b1b3
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39114
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2095584
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>