UPSTREAM: soc/amd/common/psp: Split mailbox support into v1 and v2
Family 17h redefines the PSP command and status, and therefore the
steps required to send commands via the mailbox.  Convert the existing
version into a v1 and add a v2.  New Kconfig options allow the soc to
choose v1 vs. v2.

The v2 PSP begins responding to the mailbox command when the full
bit range is written.  Define the new mailbox as a union of a u32
and a structure.

Additional PSP details may be found in the NDA publication (#55758)
  AMD Platform Security Processor BIOS Architecture Design Guide for
  AMD Family 17h Processors

Change the existing two soc functions that return pointers to void
pointers.

BUG=b:153677737

Change-Id: I4b97cdb3d02992b4b0bdea494120fe612a6b23a9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d6b7236732d8cc74545849f4b81af1d33e8758e2
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I4d358fdae07da471640856f57568059e9487f6a8
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40293
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2149459
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>