UPSTREAM: memrange: Enable memranges to support different alignments
This change enables memranges library to support addresses with
different alignments. Before this change, memranges library supported
aligning addresses to 4KiB only. Though this works for most cases, it
might not be the right alignment for every use case. Example: There
are some resource allocator changes coming up that require a different
alignment when handling the range list.

This change adds a align parameter to struct memranges that determines
the alignment of all range lists in that memrange. In order to
continue supporting current users of memranges, default alignment is
maintained as 4KiB.

BUG=b:149186922

Change-Id: I33320fcaf5ad20d4ec347f966aad8fad0b7aa0bb
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1429092d02e8157c8e6c26849e0c8aa096da9c0c
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: I1da0743ff89da734c9a0972e3c56d9f512b3d1e8
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39483
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2107468
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>