UPSTREAM: memrange: Add support for stealing required memory from given ranges
This change adds memranges_steal() which allows the user
to steal memory from the list of available ranges by providing a set
of constraints (limit, size, alignment, tag). It tries to find the
first big enough range that can satisfy the constraints, creates a
hole as per the request and returns base of the stolen memory.

BUG=b:149186922

Change-Id: If70d591a7ff60a89b905e81628985391e9c34580
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9c6274cd8fb1c9ee0eb674ce5945a05f818cb32e
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: Ibe9cfae18fc6101ab2e7e27233e45324c8117708
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39484
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2107469
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>