UPSTREAM: pciexp_device: Add option to allocate prefetch memory above 4G boundary
This change adds a Kconfig option to request allocation of prefetch
memory for hotplug devices above the 4G boundary. In order to
select this option by default and still allow users to disable this if
required, another option is added to request allocation of prefetch
memory below 4G boundary which defaults to n but can be overriden
by mainboards.

Without this change, if the number of pciexp bridges supporting
hot-plug is more than 4 or if the reserved prefetch memory size for
hot-plug cases was increased, then the resource allocator would fail
to satisfy the resource requirement below 4G boundary.

BUG=b:149186922
TEST=Enabled resource allocation above 4G for prefetch memory on volteer
and verified that it gets allocated above 4G boundary.

Change-Id: Ie3aae0e5255cba691affb692eb74c3df15f161fd
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: dcbf6454b6d2d9b3627a14126ef20ed4b9c7d954
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: I061d935eef9fcda352230b03b5cf14e467924e50
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39489
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2199207
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>