UPSTREAM: src/soc/intel/tigerlake: Fix incorrect use of Field objects in ASL
Method RAOW is assuming that the first argument is a Field object
and writing to it expecting the register to get updated. However,
the callers are passing in the value of the Field object instead.

This eventually is resulting the IMGCLK not getting enable/disabled on the
platform.

Fix this by sending the exact address of the register to be updated.

Also MCCT was setting the clock frequency in both case i.e, Clock Enable
and Disable. Split the MCCT method in two, MCON and MCOF to fix the sequencing
like below
MCON:
 Set frequency
 Enable clock
MCOF:
 Disable clock

Also, make use of MCON and MCOF methods for camera clock control in tglrvp.
This is to avoid the buildbot marking the patch unstable.

BUG=None
BRANCH=None
TEST=Build and Boot waddledoo board and verified that IMGCLKOUT for
world facing camera is enabled/disabled and able to capture images.
Build and Boot Tiger Lake RVP board and verified that IMGCLKOUT for
world facing camera is enabled/disabled and able to capture images.

Change-Id: If1e659e8059170228a5abbbaa0b589eca3cfe047
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 789bdc3d9bdd4f4a7587b8dd2dc7cc21e1f2495c
Original-Signed-off-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Original-Change-Id: I8b886255d5f38819502ae1f4af0851b5a0922b22
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39498
Original-Reviewed-by: Srinidhi N Kaushik <srinidhi.n.kaushik@intel.com>
Original-Reviewed-by: Wonkyu Kim <wonkyu.kim@intel.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2107477
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>