UPSTREAM: soc/amd/psp: Add SmmInfo command
Implement the MboxBiosCmdSmmInfo function to inform the PSP of the SoC's
SMM configuration. Once the BootDone command is sent, the PSP only
responds to commands where the buffer is in SMM memory.

Set aside a region for the core-to-PSP command buffer and the
PSP-to-core mailbox. Also add an SMM flag, which the PSP expects to read
as non-zero during an SMI.

Add calls to soc functions for the soc to populate the trigger info and
register info (v2 only).

Add functions to set up the structures needed for the SmmInfo function
in Picasso support. Issue a SW SMI, and add a new handler to call the
new PSP function.

BUG=b:153677737

Signed-off-by: Dossym Nurmukhanov <dossym@chromium.org>
Original-Commit-Id: e8ffa9ffd3cf5cb9fcade12e1f1e0dea5fc3fcf2
Original-Change-Id: I10088a53e786db788740e4b388650641339dae75
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40295
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Change-Id: I09d027b0041bd38b79a4b4e9ed0e8bcc21292a51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2155365