UPSTREAM: lp/drivers/usb: Work around QEMU XHCI register issue
The QEMU XHCI controller does not support byte/word reads from the
capability register and it expects dword reads only.

In order to make this work move the access of the capability
register fields to use macros instead of a packed struct bitfield.

This issue was filed upstream:
https://bugs.launchpad.net/qemu/+bug/1693050

The original fix attempt in 2012 was not effective:
https://github.com/qemu/qemu/commit/6ee021d41078844df60a3a466e3829a3e82776f3

With this change the controller is detected properly by the libpayload
USB drivers.

BUG=none
BRANCH=none
TEST=none

Change-Id: I6e0a48260d3686b9c4f0389cc465ba6dbc7df7c8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 287cf6c7d101d98cdbba8926651d87945534b7d9
Original-Change-Id: I048ed14921a4c9c0620c10b315b42476b6e5c512
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39838
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2149470
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>