UPSTREAM: memranges: Change align attribute to be log2 of required alignment
This change updates the align attribute of memranges to be represented
as log2 of the required alignment. This makes it consistent with how
alignment is stored in struct resource as well.

Additionally, since memranges only allow power of 2 alignments, this
change allows getting rid of checks at runtime and hence failure cases
for non-power of 2 alignments.

This change also updates the type of align to be unsigned char.

BUG=b:149186922

Change-Id: I94c72b30f2dc2154f5374d8f341a136756794ac1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1908340b6903a41750226db90d6dbd39eb527c99
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: Ie4d3868cdff55b2c7908b9b3ccd5f30a5288e62f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39810
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2126772
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>